References

SurvivalModels.Cox — Type

StatsBase.fit(Cox, @formula(Surv(T,Δ)~predictors), dataset)

arguments: 
- T: The Cox model type to fit (CoxV3)
- formula: A StatsModels.FormulaTerm specifying the survival model
- df: A DataFrame containing the variables specified in the formula

returns: 
- predictor: A Vector{String} containing the names of the predictor variables included in the model
- beta: A Vector{Float64} containing the estimated regression coefficients (β​) for each predictor
- se: A Vector{Float64} containing the standard errors of the estimated regression coefficients
- loglikelihood: A Vector{Float64} containing the log-likelihood of the fitted model. This value is repeated for each predictor row 

- coef: A vector of the estimated coefficients
- formula: The applied formula

Example: ovarian = dataset("survival", "ovarian") ovarian.FUTime = Float64.(ovarian.FUTime) (Time column needs to be Float64 type) ovarian.FUStat = Bool.(ovarian.FUStat) (Status column needs to be Bool type) model = fit(Cox, @formula(Surv(FUTime, FUStat) ~ Age + ECOG_PS), ovarian)

Types:

Cox : the base abstract type

CoxGrad<:Cox : abstract type for Cox models that are solved using gradient-based optimization

CoxLLH<:CoxGrad : abstract type for Cox models that are solved by optimizing the log-likelihood

SurvivalModels.CoxV0 — Type

CoxVO(T, Δ, X)
fit(CoxV0, @formula(Surv(T,Δ)~X), data = ...)

An implementation of the Cox proportional hazards model that minimizes the negative partial log-likelihood function (cox_nllh). This version uses the Nelder-Mead method, a derivative-free optimization algorithm.

Fields: - X::Matrix{Float64}: The design matrix of covariates, where rows correspond to individuals and columns to features - T::Vector{Float64}: The observed times, sorted in ascending order - Δ::Vector{Bool}: The event indicator vector (true for event, false for censoring)

SurvivalModels.CoxV1 — Type

CoxV1(T, Δ, X)
fit(CoxV1, @formula(Surv(T,Δ)~X), data = ...)

The first implementation of the Cox proportional hazards model uses optimization libraries (Optimization.jl, Optim.jl) for coefficient estimation. It uses the BFGS algorithm to minimize the negative partial log-likelihood.

Fields: - X::Matrix{Float64}: The design matrix of covariates, where rows correspond to individuals and columns to features. - T::Vector{Float64}: The observed times, sorted in ascending order - Δ::Vector{Int64}: The event indicator vector (true for event, false for censoring)

SurvivalModels.CoxV2 — Type

CoxV2(T, Δ, X)
fit(CoxV2, @formula(Surv(T,Δ)~X), data = ...)

The second implementation of the Cox proportional hazards model uses a Newton-Raphson-like iterative update that directly calculates and utilizes the gradient and Hessian matrix. This version is updating coefficients via the update! function.

Fields: - X::Matrix{Float64}: The design matrix of covariates, where rows correspond to individuals and columns to features - T::Vector{Float64}: The observed times, sorted in ascending order - Δ::Vector{Int64}: The event indicator vector (true for event, false for censoring) - R::BitMatrix: A boolean risk matrix, where 'R[i,j]' is 'true' if individual 'j' is at risk at time 'T[i]'

SurvivalModels.CoxV3 — Type

CoxV3(T, Δ, X)
fit(CoxV3, @formula(Surv(T,Δ)~X), data = ...)
fit(Cox, @formula(Surv(T,Δ)~X), data = ...)

The third implementation of the Cox proportional hazards model represents a highly optimized and significantly faster iteration compared to previous implementation, CoxV2.

This is the default implementation called when you ask for a Cox model.

Fields: - Xᵗ::Matrix{Float64}: The design matrix of covariates, transposed (m rows, n columns) - sX::Vector{Float64}: Sum of X' multiplied by Δ - T::Vector{Float64}: The observed times sorted in descending order - Δ::Vector{Bool}: The event indicator vector (true for event, false for censoring) - loss::Vector{Float64}: Stores the current negative partial log-likelihood value - G::Vector{Float64}: Stores the gradient vector - H::Matrix{Float64}: Stores the Hessian matrix - S₁::Vector{Float64}: Sum of rⱼxₖⱼ - S₂::Matrix{Float64}: Sum of rⱼxₖⱼ * xⱼ - μ::Vector{Float64}: Updates the gradient and Hessian - η::Vector{Float64}: ηi = Xiβ - r::Vector{Float64}: ri = exp(ηi) - R::Vector{UnitRange{Int64}}: A vector of the risk ranges for each output time.

SurvivalModels.CoxV4 — Type

CoxV4(T, Δ, X)
fit(CoxV4, @formula(Surv(T,Δ)~X), data = ...)

The fourth implementation of the Cox proportional hazards model uses Hessian approximation based on a pre-calculated estimation. This version was created for when it might be difficult to work with full Hessian , offering faster iterations by using a Hessian approximation.

Fields: - X::Matrix{Float64}: The design matrix of covariates, where rows correspond to individuals and columns to features - T::Vector{Float64}: The observed times sorted in ascending order - Δ::Vector{Bool}: The event indicator vector (true for event, false for censoring) - sX::Vector{Float64}: Sum of X' multiplied by Δ - G::Vector{Float64}: Stores the gradient vector - η::Vector{Float64}: ηi = Xiβ - A::Vector{Float64}: Ai = exp(ηi) - B::Vector{Float64}: Stores the majoration elements of the Hessian matrix - C::Vector{Float64}: Used in the mkA! function - K::Vector{Int64}: Number of events at each unique observed event time - loss::Vector{Float64}: Stores the current negative partial log-likelihood value, used in CoxLLH getβ

SurvivalModels.CoxV5 — Type

CoxV5(T, Δ, X)
fit(CoxV5, @formula(Surv(T,Δ)~X), data = ...)

The fifth implementation of the Cox proportional hazards model uses a pre-calculated Hessian approximation for faster iterations, like CoxV4. Its approach is similar to CoxV3.

Fields: - Xᵗ::Matrix{Float64}: The design matrix of covariates, transposed (m rows, n columns) - sX::Vector{Float64}: Sum of X' multiplied by Δ - T::Vector{Float64}: The observed times sorted in descending order - Δ::Vector{Bool}: The event indicator vector (true for event, false for censoring) - loss::Vector{Float64}: Stores the current negative partial log-likelihood value - G::Vector{Float64}: Stores the gradient vector. - S₁::Vector{Float64}: Sum of rⱼxₖⱼ - μ::Vector{Float64}: Currently unused in update! function - η::Vector{Float64}: ηi = Xiβ - r::Vector{Float64}: ri = exp(ηi) - R::Vector{UnitRange{Int64}}: - B::Vector{Float64}: Stores the majoration elements of the Hessian matrix

[1]

D. R. Cox. Regression models and life-tables. Journal of the Royal Statistical Society: Series B (Methodological) 34, 187–202 (1972).