Cox models

one ref: [1].

Theory

The Cox Proportional Hazards Model is a semi-parametric model used to analyze time-to-event data. It models the relationship between the survival time of an individual and a set of covariates. It is defined by the hazard function:

\[h(t | \mathbf{X}) = h_0(t) \exp(\mathbf{X}^T\mathbf{\beta})\]

where: $h_0(t)$ is the baseline hazard function, $\mathbf{X}$ is the covariate vector, $\mathbf{\beta}$ is the vector of regression coefficients.

The partial-likelihood function for the Cox model is given by:

\[L(\mathbf{\beta}) = \prod_{j=1}^{k} \frac{\prod_{ \Delta_i=1} \exp(\mathbf{X}_i^T\mathbf{\beta})}{\left( \sum_{l \in R_j} \exp(\mathbf{X}_l^T\mathbf{\beta}) \right)}\]

Our goal is to maximize the log-partial-likelihood or, equivalently, to minimize its negative, which we define as our loss function:

\[ \text{Loss}(\mathbf{\beta}) = - \log L(\mathbf{\beta}) = - \sum_{j=1}^{k} \left( \sum \mathbf{X}_i^T\mathbf{\beta} - \log \left( \sum_{l \in R_j} \exp(\mathbf{X}_l^T\mathbf{\beta}) \right) \right) \]

The loss function is coded as follows:

function loss(beta, M::Cox)
    η = M.X*beta
    return dot(M.Δ, log.((M.T .<= M.T') * exp.(η)) .- η)
end

We differentiate once and obtain the gradient:

\[ \frac{\partial}{\partial \beta_k} \text{Loss}(\mathbf{\beta}) = - \sum_{i=1}^{n} \left( X_{ik} - \frac{\sum_{j \in R_i} \exp(\mathbf{\beta}^T\mathbf{X}_j) X_{jk}}{\sum_{j \in R_i} \exp(\mathbf{\beta}^T\mathbf{X}_j)} \right) \]

And the Hesian matrix:

\[ \frac{\partial^2}{\partial \beta_k \partial \beta_l} \text{Loss}(\mathbf{\beta}) = \sum_{i=1}^{n} \left[ \frac{\sum_{j \in R_i} \exp(\mathbf{\beta}^T\mathbf{X}_j) X_{jk}X_{jl}}{\sum_{j \in R_i} \exp(\mathbf{\beta}^T\mathbf{X}_j)} - \frac{\left( \sum_{j \in R_i} \exp(\mathbf{\beta}^T\mathbf{X}_j) X_{jk} \right) \left( \sum_{j \in R_i} \exp(\mathbf{\beta}^T\mathbf{X}_j) X_{jl} \right)}{\left( \sum_{j \in R_i} \exp(\mathbf{\beta}^T\mathbf{X}_j) \right)^2} \right] \]

The information matrix is defined as the negative of the Hessian matrix of the log-likelihood function, evaluated at the point of the estimated coefficients.

\[ I(\hat{\boldsymbol{\beta}}) = -H(\hat{\boldsymbol{\beta}}) \]

But, in the earlier formula, $\mathbf{H}_{\text{Loss}}$​ was for Loss(β), which is -log L(β) $\mathbf{H}_{\text{Loss}} = - \mathbf{H}_{\text{log-likelihood}}$. Therefore, the observed Information Matrix is equal to $\mathbf{H}_{\text{Loss}}$ itself.

The variance (and covariance) of our estimators $\hat{\boldsymbol{\beta}}$ are obtained by inverting the observed information matrix.

\[\text{Var}(\hat{\boldsymbol{\beta}}) = I(\hat{\boldsymbol{\beta}})^{-1}\]

This final matrix contains:

On its diagonal: the variances of each coefficient ($\text{Var}(\hat{\beta}_1)$, $\text{Var}(\hat{\beta}_2)$, ...).

Off-diagonal: the covariances between pairs of coefficients.

The standard error for a specific coefficient ($\hat{\beta}_k$) is the square root of its variance.

\[SE(\hat{\beta}_k) = \sqrt{\text{Var}(\hat{\beta}_k)}\]

To determine if a variable has a statistically significant effect, a Wald test is performed. A z-score is calculated: $z = \frac{\text{Coefficient}}{\text{Erreur Type}} = \frac{\hat{\beta}}{SE(\hat{\beta})}$

This $z$-score is then compared to a normal distribution to obtain a $p$-value. A low $p$-value (typically < 0.05) suggests that the coefficient is significantly different from zero.

The p-value for each coefficient is calculated by comparing its z-score to a standard normal distribution. This p-value indicates the probability of observing a z-score as extreme as, or more extreme than, the one calculated, assuming the null hypothesis (that the coefficient is zero) is true.

The standard error allows for the construction of a confidence interval (CI) around the coefficient, which provides a range of plausible values for the true coefficient.

The general formula for a $(1 - \alpha) \times 100\%$ confidence interval is:

\[\text{IC pour } \hat{\beta} = \hat{\beta} \pm z_{\alpha/2} \times SE(\hat{\beta})\]

Let us see for example the output on the colon dataset:

using SurvivalModels
using RDatasets

# ovarian = dataset("survival", "ovarian")
# ovarian.FUTime = Float64.(ovarian.FUTime)
# ovarian.FUStat = Bool.(ovarian.FUStat)
# model = fit(Cox, @formula(Surv(FUTime, FUStat) ~ Age + ECOG_PS), ovarian)

colon = dataset("survival", "colon")
colon.Time = Float64.(colon.Time)
colon.Status = Bool.(colon.Status)
model_colon = fit(Cox, @formula(Surv(Time, Status) ~ Age + Rx), colon)

3×5 DataFrame

Row	Predictor	β	SE	P_Value	z

String	Float64	Float64	Float64	Float64

1	Age	-0.00205614	0.00280669	0.463812	-0.732585

2	Rx: Lev	-0.0200488	0.0768372	0.79415	-0.260925

3	Rx: Lev+5FU	-0.439289	0.0839364	1.66245e-7	-5.23359

The outputed datafrae contains collumns with respectively the name of the predictor, the obtained coefficients, its standard error, the associated p-value and the test statistic z as just described.

SurvivalModels.Cox — Type

StatsBase.fit(Cox, @formula(Surv(T,Δ)~predictors), dataset)

arguments: 
- T: The Cox model type to fit (CoxV3)
- formula: A StatsModels.FormulaTerm specifying the survival model
- df: A DataFrame containing the variables specified in the formula

returns: 
- predictor: A Vector{String} containing the names of the predictor variables included in the model
- beta: A Vector{Float64} containing the estimated regression coefficients (β​) for each predictor
- se: A Vector{Float64} containing the standard errors of the estimated regression coefficients
- loglikelihood: A Vector{Float64} containing the log-likelihood of the fitted model. This value is repeated for each predictor row 

- coef: A vector of the estimated coefficients
- formula: The applied formula

Example: ovarian = dataset("survival", "ovarian") ovarian.FUTime = Float64.(ovarian.FUTime) (Time column needs to be Float64 type) ovarian.FUStat = Bool.(ovarian.FUStat) (Status column needs to be Bool type) model = fit(Cox, @formula(Surv(FUTime, FUStat) ~ Age + ECOG_PS), ovarian)

Types:

Cox : the base abstract type

CoxGrad<:Cox : abstract type for Cox models that are solved using gradient-based optimization

CoxLLH<:CoxGrad : abstract type for Cox models that are solved by optimizing the log-likelihood

Different versions of the optimisation routine

To implement the Cox proportional hazards model, different versions were coded, using different methods. The final goal is to compare these versions and choose the most efficient one: the fastest and the closest to the true values of the coefficients.

V0

SurvivalModels.CoxV0 — Type

CoxVO(T, Δ, X)
fit(CoxV0, @formula(Surv(T,Δ)~X), data = ...)

An implementation of the Cox proportional hazards model that minimizes the negative partial log-likelihood function (cox_nllh). This version uses the Nelder-Mead method, a derivative-free optimization algorithm.

Fields: - X::Matrix{Float64}: The design matrix of covariates, where rows correspond to individuals and columns to features - T::Vector{Float64}: The observed times, sorted in ascending order - Δ::Vector{Bool}: The event indicator vector (true for event, false for censoring)

V1: Implementation using the 'Optimization.jl' Julia package

SurvivalModels.CoxV1 — Type

CoxV1(T, Δ, X)
fit(CoxV1, @formula(Surv(T,Δ)~X), data = ...)

The first implementation of the Cox proportional hazards model uses optimization libraries (Optimization.jl, Optim.jl) for coefficient estimation. It uses the BFGS algorithm to minimize the negative partial log-likelihood.

Fields: - X::Matrix{Float64}: The design matrix of covariates, where rows correspond to individuals and columns to features. - T::Vector{Float64}: The observed times, sorted in ascending order - Δ::Vector{Int64}: The event indicator vector (true for event, false for censoring)

V2: Implementation using the gradient and the Hessian matrix

SurvivalModels.CoxV2 — Type

CoxV2(T, Δ, X)
fit(CoxV2, @formula(Surv(T,Δ)~X), data = ...)

The second implementation of the Cox proportional hazards model uses a Newton-Raphson-like iterative update that directly calculates and utilizes the gradient and Hessian matrix. This version is updating coefficients via the update! function.

Fields: - X::Matrix{Float64}: The design matrix of covariates, where rows correspond to individuals and columns to features - T::Vector{Float64}: The observed times, sorted in ascending order - Δ::Vector{Int64}: The event indicator vector (true for event, false for censoring) - R::BitMatrix: A boolean risk matrix, where 'R[i,j]' is 'true' if individual 'j' is at risk at time 'T[i]'

V3: Improved version of V2 (much faster) because non-allocative.

SurvivalModels.CoxV3 — Type

CoxV3(T, Δ, X)
fit(CoxV3, @formula(Surv(T,Δ)~X), data = ...)
fit(Cox, @formula(Surv(T,Δ)~X), data = ...)

The third implementation of the Cox proportional hazards model represents a highly optimized and significantly faster iteration compared to previous implementation, CoxV2.

This is the default implementation called when you ask for a Cox model.

Fields: - Xᵗ::Matrix{Float64}: The design matrix of covariates, transposed (m rows, n columns) - sX::Vector{Float64}: Sum of X' multiplied by Δ - T::Vector{Float64}: The observed times sorted in descending order - Δ::Vector{Bool}: The event indicator vector (true for event, false for censoring) - loss::Vector{Float64}: Stores the current negative partial log-likelihood value - G::Vector{Float64}: Stores the gradient vector - H::Matrix{Float64}: Stores the Hessian matrix - S₁::Vector{Float64}: Sum of rⱼxₖⱼ - S₂::Matrix{Float64}: Sum of rⱼxₖⱼ * xⱼ - μ::Vector{Float64}: Updates the gradient and Hessian - η::Vector{Float64}: ηi = Xiβ - r::Vector{Float64}: ri = exp(ηi) - R::Vector{UnitRange{Int64}}: A vector of the risk ranges for each output time.

V4: Majoration of the Hessian matrix by a universal bound.

SurvivalModels.CoxV4 — Type

CoxV4(T, Δ, X)
fit(CoxV4, @formula(Surv(T,Δ)~X), data = ...)

The fourth implementation of the Cox proportional hazards model uses Hessian approximation based on a pre-calculated estimation. This version was created for when it might be difficult to work with full Hessian , offering faster iterations by using a Hessian approximation.

Fields: - X::Matrix{Float64}: The design matrix of covariates, where rows correspond to individuals and columns to features - T::Vector{Float64}: The observed times sorted in ascending order - Δ::Vector{Bool}: The event indicator vector (true for event, false for censoring) - sX::Vector{Float64}: Sum of X' multiplied by Δ - G::Vector{Float64}: Stores the gradient vector - η::Vector{Float64}: ηi = Xiβ - A::Vector{Float64}: Ai = exp(ηi) - B::Vector{Float64}: Stores the majoration elements of the Hessian matrix - C::Vector{Float64}: Used in the mkA! function - K::Vector{Int64}: Number of events at each unique observed event time - loss::Vector{Float64}: Stores the current negative partial log-likelihood value, used in CoxLLH getβ

V5

SurvivalModels.CoxV5 — Type

CoxV5(T, Δ, X)
fit(CoxV5, @formula(Surv(T,Δ)~X), data = ...)

The fifth implementation of the Cox proportional hazards model uses a pre-calculated Hessian approximation for faster iterations, like CoxV4. Its approach is similar to CoxV3.

Fields: - Xᵗ::Matrix{Float64}: The design matrix of covariates, transposed (m rows, n columns) - sX::Vector{Float64}: Sum of X' multiplied by Δ - T::Vector{Float64}: The observed times sorted in descending order - Δ::Vector{Bool}: The event indicator vector (true for event, false for censoring) - loss::Vector{Float64}: Stores the current negative partial log-likelihood value - G::Vector{Float64}: Stores the gradient vector. - S₁::Vector{Float64}: Sum of rⱼxₖⱼ - μ::Vector{Float64}: Currently unused in update! function - η::Vector{Float64}: ηi = Xiβ - r::Vector{Float64}: ri = exp(ηi) - R::Vector{UnitRange{Int64}}: - B::Vector{Float64}: Stores the majoration elements of the Hessian matrix

Comparison of the different methods speed

We propose to compare the different methods on simulated data, with varying number of lines and columns, to verify empirically the theoretical complexity of the different methods. We will then compare the results with Julia's and R's existing Cox implementation.

using SurvivalModels, Plots, Random, Distributions, StatsBase, LinearAlgebra, DataFrames, RCall, Survival

using SurvivalModels: getβ, CoxV0, CoxV1, CoxV2, CoxV3, CoxV4, CoxV5

struct CoxVJ
    T::Vector{Float64}
    Δ::Vector{Bool}
    X::Matrix{Float64}
    function CoxVJ(T,Δ,X)
        new(T,Bool.(Δ),X)
    end
end

function SurvivalModels.getβ(M::CoxVJ)
    return fit(Survival.CoxModel, M.X, Survival.EventTime.(M.T,M.Δ)).β
end

R"""
library(survival)
"""

struct CoxVR
    df::DataFrame
    function CoxVR(T,Δ,X)
        df = DataFrame(X,:auto)
        df.status = Δ
        df.time = T
        new(df)
    end
end

function SurvivalModels.getβ(M::CoxVR)
    df = M.df
    @rput df
    R"""
    beta  <- coxph(Surv(time,status)~., data = df, ties="breslow")$coefficients
    """
    @rget beta
    return beta
end

# Creating a dictionary for all the models:
# Label => (constructor, plotting color)

const design = Dict(
"V0"=> (CoxV0, :blue),
"V1"=> (CoxV1, :orange),
"V2"=> (CoxV2, :brown),
"V3"=> (CoxV3, :purple),
"V4"=> (CoxV4, :green),
"V5"=> (CoxV5, :yellow),
"VR"=> (CoxVR, :red),
"VJ"=> (CoxVJ, :black)
);

# Function to simulate data for different rows and column numbers (n max = 2000 et m max = 20):
function simulate_survival_data(n, m; censor_rate = 0.2, β=randn(m))
    Random.seed!(42)
    X = hcat(
        [randn(n)       for _ in 1:cld(m,3)]..., # about 1/3
        [rand(n)        for _ in 1:cld(m,3)]..., # about 1/3
        [exp.(randn(n)) for _ in 1:(m-2cld(m,3))]... # the rest.
    )
    η = X * β
    λ₀ = 1
    U = rand(n)
    O = -log.(U) ./ (λ₀ .* exp.(η))
    lc = quantile(O, 1 - censor_rate)
    C = rand(Exponential(lc), n)
    T = min.(O, C)
    Δ = Bool.(T .<= C)
    return (T, Δ, X)
end

# Run the models and get the running time, the β coefficients and the difference between the true β and the obtained ones:
function run_models()
    Ns = (500, 1000, 2000)
    Ms = (10, 20)
    true_betas = randn(maximum(Ms))
    df = []
    for n in Ns, m in Ms
        if (n == 2000) | (m == 20) # Only if they end up in the graphs.
            data = simulate_survival_data(n,m, β = true_betas[1:m])
            for (name, (constructor, _)) in design
                display((n,m,name))
                model = constructor(data...)
                beta = getβ(model)
                time = @elapsed getβ(model)
                push!(df, (
                    n = n,
                    m = m,
                    name = name,
                    time = time,
                    beta = beta,
                    diff_to_truth = sqrt(sum((beta .- true_betas[1:m]).^2)/sum(true_betas[1:m].^2)),
                ))
            end
        end
    end
    df = DataFrame(df)
    sort!(df, :name)
    return df
end

# Plot the results, starting with the time:
function timing_graph(df)
    group1 = groupby(filter(r -> r.m==20, df), :name)
    p1 = plot(; xlabel = "Number of observations (n)",
        ylabel = "Time (in seconds)",
        yscale= :log10,
        xscale= :log10,
        title = "For m=20 covs., varying n",
        legend = :bottomright,
        lw = 1);
    for g in group1
        plot!(p1, g.n, g.time, label = g.name[1] , color = design[g.name[1]][2], marker = :circle, markersize = 3)
    end
    group2 = groupby(filter(r -> r.n==2000, df), :name)
        p2 = plot(; xlabel = "Number of covariates (m)",
            ylabel = "Temps (ms)",
            yscale= :log10,
            xscale= :log10,
            title = "For n=2000 obs., varying m",
            legend = :bottomright,
            lw = 1);
        for g in group2
            plot!(p2, g.m, g.time, label = g.name[1] , color = design[g.name[1]][2], marker = :circle, markersize = 3)
        end
        p = plot(p1,p2, size=(1200,600), plot_title = "Runtime (logscale) of the various implementations")
        return p
end

timing_graph (generic function with 1 method)

x=1
# df = run_models()
# timing_graph(df)

1

comments on the graph.

A zoom on our implementation vs Survival.jl vs R::survival:

x=1
# timing_graph(filter(r -> r.name ∈ ("V3", "VJ", "VR"), df))

1

So we are about x10 faster than the reference implmentation of R (and than the previous Julia attemps) on this example.

function beta_correctness_graphs(df; ref="VJ")

    reflines = filter(r -> r.name == ref, df)
    rename!(reflines, :beta => :refbeta)
    select!(reflines, Not([:name, :time, :diff_to_truth]))
    otherlines = filter(r -> r.name != ref, df)
    rez = leftjoin(otherlines, reflines, on=[:n,:m])
    percent(x,y) = sqrt(sum((x .- y).^2)/sum(y .^2))
    rez.error = percent.(rez.beta, rez.refbeta)
    select!(rez, [:n,:m,:name,:error])
    rez = filter!(r -> !isnan(r.error), rez)

    group1 = groupby(filter(r -> r.m==20, rez), :name)
    p1 = plot(; xlabel = "Number of observations (n)",
                ylabel = "L2dist to $ref's β",
                yscale=:log10,
                xscale= :log10,
                title = "m=20, varying n",
                legend = :bottomright,
                lw = 1);
    for g in group1
        plot!(p1, g.n, g.error, label = g.name[1] , color = design[g.name[1]][2], marker = :circle, markersize = 3)
    end

    group2 = groupby(filter(r -> r.n==2000, rez), :name)
    p2 = plot(; xlabel = "Nomber of covariates (m)",
                ylabel = "L2Dist to $ref's β",
                yscale=:log10,
                xscale= :log10,
                title = "n=2000, varying m",
                legend = :bottomright,
                lw = 1);
    for g in group2
        plot!(p2, g.m, g.error, label = g.name[1] , color = design[g.name[1]][2], marker = :circle, markersize = 3)
    end
    p = plot(p1,p2, size=(1200,600), plot_title="β-correctness w.r.t. $ref's version.")
    return p
end

# beta_correctness_graphs(df)

beta_correctness_graphs (generic function with 1 method)

function beta_wrt_truth(df)
    group1 = groupby(filter(r -> r.m==20, df), :name)
    p1 = plot(; xlabel = "Number of observations (n)",
                ylabel = "L2dist to the truth",
                yscale=:log10,
                xscale= :log10,
                title = "m=20, varying n",
                legend = :bottomright,
                lw = 1);
    for g in group1
        plot!(p1, g.n, g.diff_to_truth, label = g.name[1] , color = design[g.name[1]][2], marker = :circle, markersize = 3)
    end

    group2 = groupby(filter(r -> r.n==2000, df), :name)
    p2 = plot(; xlabel = "Nomber of covariates (m)",
                ylabel = "L2Dist to the truth",
                yscale=:log10,
                xscale= :log10,
                title = "n=2000, varying m",
                legend = :bottomright,
                lw = 1);
    for g in group2
        plot!(p2, g.m, g.diff_to_truth, label = g.name[1] , color = design[g.name[1]][2], marker = :circle, markersize = 3)
    end
    p = plot(p1,p2, size=(1200,600), plot_title="β-correctness w.r.t. the truth.")
    return p
end

# beta_wrt_truth(df)

beta_wrt_truth (generic function with 1 method)

[1]

D. R. Cox. Regression models and life-tables. Journal of the Royal Statistical Society: Series B (Methodological) 34, 187–202 (1972).